# Cache for GetGraphTypeFromType