--- a/src/GraphQL/AssemblyInfo.cs
+++ b/src/GraphQL/AssemblyInfo.cs
@@ -1,1 +1,2 @@
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("GraphQL.Tests")]
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("GraphQL.DataLoader.Tests")]