fixes #2102 

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.16299.726 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i7 CPU 920 2.67GHz (Nehalem), 1 CPU, 8 logical and 4 physical cores
Frequency=2607424 Hz, Resolution=383.5203 ns, Timer=TSC
.NET Core SDK=5.0.101
  [Host]     : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT
  DefaultJob : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT


```

before

|        Method | UseCaching |         Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |----------- |-------------:|----------:|----------:|---------:|--------:|------:|----------:|
| **Introspection** |      **False** | **2,150.563 μs** | **4.3997 μs** | **4.1155 μs** | **128.9063** | **23.4375** |     **-** | **544.48 KB** |
|          Hero |      False |    41.554 μs | 0.0672 μs | 0.0561 μs |   2.1362 |       - |     - |   8.81 KB |
| **Introspection** |       **True** | **1,685.397 μs** | **2.7979 μs** | **2.4802 μs** | **121.0938** | **33.2031** |     **-** | **504.74 KB** |
|          Hero |       True |     9.085 μs | 0.0126 μs | 0.0118 μs |   1.0071 |       - |     - |   4.14 KB |

after

|        Method | UseCaching |         Mean |     Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------- |----------- |-------------:|----------:|----------:|---------:|--------:|------:|----------:|
| **Introspection** |      **False** | **2,150.850 μs** | **2.4876 μs** | **2.3269 μs** | **132.8125** | **31.2500** |     **-** | **544.45 KB** |
|          Hero |      False |    41.682 μs | 0.0451 μs | 0.0422 μs |   2.1362 |       - |     - |   8.81 KB |
| **Introspection** |       **True** | **1,669.312 μs** | **2.4231 μs** | **2.2666 μs** | **119.1406** | **39.0625** |     **-** | **504.76 KB** |
|          Hero |       True |     9.343 μs | 0.0143 μs | 0.0127 μs |   1.0071 |       - |     - |   4.14 KB |

😕 

Also I suggest to make `TypeExtensions` internal.