@sungam3r 2021-01-17
Perhaps my machine is getting inaccurate results. Anyway, if there is a difference in speed, then it is minimal.