@Shane32 2021-01-18
When is GetGraphTypeFromType used during execution?