@Shane32 2021-01-18
GetGraphTypeFromType seems pretty important to keep public.  Not sure, but I might be using that in some of my projects.  These also I would keep public:
* IsGraphType
* GraphQLName
* GetEnumerableElementType

Probably these too:
* Description
* ObsoleteMessage

The rest are questionable, although I would certainly remove `As` altogether as it is an alias for a native C# keyword.

p.s. IsGraphType should be rewritten with IsAssignableFrom